+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Note that since we don't clearly distinguish between a public and private interf
 - Support measurements for coarse models
 - Fix `element-point` visual not using child unit
 - Ignore `renderables` with empty draw count
+- Add experimental support for `esbuild` for development.
+  - Use `npm run dev` for faster development builds
+
 
 ## [v4.12.0] - 2025-02-28